backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36
pydub==0.25.1
librosa==0.10.1
openai-whisper>=20231117
nltk==3.8.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9; platform_system != "Windows"
pytest==7.4.3
pillow>=10.2.0
